+++ py7zip/py7zip.py
@@ -9,14 +9,17 @@ class Py7zip:
     """ An unofficial, cross-platform, lightweight, and easy-to-use wrapper
         for 7zip command-line binaries (7za) in Python. """
 
-    def __init__(self):
+    def __init__(self, verbose=False, debug=False):
         """ Initializes the class and variables. """
+        self.verbose = verbose
+        self.debug = debug
+
         self.platform = platform.system().lower()
         self.supported = ['windows', 'linux', 'darwin']
         self.architecture = platform.architecture()[0]
         self.username = 'aliasfoxkde'
         self.app_name = 'py7zip'
-        self.base_bin_url = f'https://github.com/{self.username}/{self.app_name}/tree/main/bin'
+        self.base_bin_url = f'https://github.com/{self.username}/{self.app_name}/raw/main/bin/'
         self.debug_info = platform.uname()
 
         if 'AMD64' in platform.machine():
@@ -41,6 +44,10 @@ class Py7zip:
         except Exception as err:
             print(err)
 
+    @staticmethod
+    def cd(directory_path):
+        os.chdir(directory_path)
+
     def download_binary(self):
         """ Dynamically downloads the binary required for the given platform. """
         binary_url = self.get_binary_url()
@@ -53,42 +60,61 @@ class Py7zip:
 
     def get_binary_url(self):
         """ Dynamically gets the URL of the required system binary. """
-        # Useful commands: platform: .system(), .release(), platform.version()
         if self.platform not in self.supported:
             raise NotImplementedError(f"Platform '{self.platform}' is not supported.")
         return self.url
 
+    def wrapper(self, src, dst, options='', method="decompress"):
+        """ Method used to extract an archive. """
+        if method == "compress":
+            command = f'"7za{self.extension}" a "{dst}" "{src}" {options}'
+            success_msg = f"Backup created from '{src}' to '{dst}'."
+            error_msg = f"Failed to extract archive from '{src}' to '{dst}'."
+        else:  # "compress" method as default
+            command = f'"7za{self.extension}" x "{src}" -o"{dst}" {options}'
+            success_msg = f"Extracted archive from '{src}' to '{dst}'"
+            error_msg = f"Failed to create backup from '{src}' to '{dst}'."
+
+        try:
+            result = subprocess.run(command, capture_output=True, text=True, shell=True, check=True)
+            if self.debug:
+                print(result.stdout)
+            if self.verbose:
+                print(success_msg)
+        except subprocess.CalledProcessError as e:
+            if self.debug:
+                print(f"Error: {e.stderr}.")
+            if self.verbose:
+                print(error_msg)
+
+    def decompress(self, src, dst, options=''):
+        """ Method used to extract an archive. """
+        self.wrapper(src, dst, options='', method="decompress")
+
+    def extract(self, src, dst, options=''):
+        """ Alias used for 'compress' method. """
+        self.wrapper(src, dst, options='', method="decompress")
+
+    def compress(self, src, dst, options=''):
+        """ Method used to backup files, folders, or directories to an archive. """
+        self.wrapper(src, dst, options='', method="compress")
+
+    def archive(self, src, dst, options=''):
+        """ Alias used for decompress method. """
+        self.wrapper(src, dst, options='', method="compress")
+
+    def backup(self, src, dst, options=''):
+        """ Alias used for decompress method. """
+        self.wrapper(src, dst, options='', method="compress")
+
+    def full(self, src, dst, options=''): pass
+    def incremental(self, src, dst, options=''): pass
+    def differential(self, src, dst, options=''): pass
+
     def snapshot(self, src, dst, options=''):
         """ Streamlines the process for creating a snapshot. """
         pass
-        
-    def extract(self, src, dst, options=''):
-        """ Method used to extract an archive. """
-        command = f'"{self.binary_path}" x "{src}" -o"{dst}" {options}'
-        result = subprocess.run(command, capture_output=True, text=True, shell=True)
-        print(result.stdout)
-        if result.returncode == 0:
-            print(f"Extracted archive from {src} to {dst}")
-        else:
-            print(f"Failed to extract archive from {src} to {dst}. Error: {result.stderr}")
-
-    def backup(self, src, dst, options=''):
-        """ Method used to backup files, folders, or directories to an archive. """
-        def full(self, src, dst, options=''): pass
-        def incremental(self, src, dst, options=''): pass
-        def differential(self, src, dst, options=''): pass
-        
-        command = f'"{self.binary_path}" a "{dst}" "{src}" {options}'
-        result = subprocess.run(command, capture_output=True, text=True, shell=True)
-        print(result.stdout)
-        if result.returncode == 0:
-            print(f"Backup created from {src} to {dst}")
-        else:
-            print(f"Failed to create backup from {src} to {dst}. Error: {result.stderr}")
 
 
 if __name__ == "__main__":
-    # Development Testing
-    py7z = Py7zip()
-    print(py7z.get_binary_url())
     print("This module is intended to be imported, not run directly!")